+++ CHANGELOG.md
@@ -1,3 +1,29 @@
+## [3.0.0] - 2022-04-29
+
+* New arguments for Nylo's `init` method: setup, setupFinished
+* Add init method for NyState class
+* New helpers: nyEvent, nyApi and Backpack
+* assert condition added to DefaultResponse class
+
+## [2.8.0] - 2022-04-21
+
+* Revert new init method in Nylo
+
+## [2.7.0] - 2022-04-21
+
+* Version bump
+
+## [2.6.1] - 2022-04-21
+
+* Fix Nylo init method with router param
+
+## [2.6.0] - 2022-04-21
+
+* Revert init change in NyState.
+* New Metro command to create events in Nylo.
+* bootApplication helper added.
+* Small refactor to folder names
+
 ## [2.5.0] - 2022-04-19
 
 * New Metro command to create Providers in Nylo.